If the argument list is empty, call the non- varargs overload of the
* tools/m4/ctor.m4: If the argument list is empty, call the non-
varargs overload of the Glib::ConstructParams constructor instead
of using an empty varargs list.  This mistake was exposed thanks
to the recent addition of G_GNUC_NULL_TERMINATED to the varargs
constructor declaration.  Fortunately it was harmless in this case.
Also use the opportunity to clean up the M4 code and prefix builtin
M4 macros with m4_, so that we may get rid of the unprefixed macros
altogether some day.

svn path=/trunk/; revision=376