Merged from glibmm-2-14: 2007-12-28  Murray Cumming  <murrayc@murrayc.com>
* tools/m4/class_gobject.m4:
* tools/m4/ctor.m4:
_INITIALLY_UNOWNED_SINK: Rename from INITIALLY_UNOWNED_SINK, 
for consistency, because this could be used directly from .ccg files 
that use _CONSTRUCT().
Remove the debug code for the else case.

2007-12-28  Murray Cumming  <murrayc@murrayc.com>

* tools/m4/class_gobject.m4:
* tools/m4/ctor.m4:
Added the _DERIVES_INITIALLY_UNOWNED macro, for use 
after _CLASS_GOBJECT(), if the C type derives 
from GInitiallyUnowned rather than just GObject.
(Not for use with GtkObject-derived classes).
This does g_object_ref_sink() so that we get a normally-behaved 
GObject.

svn path=/trunk/; revision=479