Added _DO_NOT_DERIVE_GTYPE for use in .hg files when we do not want to
2008-01-25  Murray Cumming  <murrayc@murrayc.com>

* tools/m4/class_shared.m4: Added _DO_NOT_DERIVE_GTYPE for use in .hg 
files when we do not want to derive a new GType, meaning of course that 
we cannot have default signal handlers or vfuncs.
* gio/src/volumemonitor.hg: Use _DO_NOT_DERIVE_GTYPE, because 
gio tries to call every type that derives from G_TYPE_VOLUME_MONITOR, 
which seems unwise to me - bug #511814.
This makes gtkmm-documentation/examples/book/volumes not crash.

svn path=/trunk/; revision=540