Allow _CLASS_INTERFACE() to take two extra optional paremeters to specify
2008-02-04  Murray Cumming  <murrayc@murrayc.com>

* tools/m4/class_interface.m4: Allow _CLASS_INTERFACE() to take two 
extra optional paremeters to specify the base class, so that appropriate 
code is generated when we derive from another Interface - which seems 
appropriate when the GInterface has a prerequisite.
* gio/src/loadableicon.ccg:
* gio/src/loadableicon.hg: Derive from Icon, because it is a prerequisite 
of this interface. Use the extra _CLASS_INTERFACE() parameters.
* gio/src/fileicon.hg: Do not derive from Icon because that happens 
through LoadableIcon now. But do use _IMPLEMENTS_INTERFACE(Icon) here, 
because only a Glib::Object can actually implement an interface.

svn path=/trunk/; revision=580