Make the callbacks take a Tree<> instead of just the data, so they can use
2008-07-29  Szilárd Pfeiffer  <szilard.pfeiffer@gmail.com>

* glib/src/tree.hg: Make the callbacks take a Tree<> instead of just 
the data, so they can use methods on the tree (which can be a node 
in the tree).
gobject_: Make this protected.
Provide the this pointer as data to g_node_new() so we can retrieve 
it later.
Removed children_ and parent_ because we don't need a separate store now that 
we can get the C++ instance from the gobject instance.
owns_gobject_: Removed because it is was always true, so the gobject was 
always destroyed (and still is).
* tests/glibmm_tree/main.cc: Updated for the changed API.
Bug #520778.

svn path=/trunk/; revision=697