Work around Automake cleverness failure
* gio/giomm/filelist.am (giomm_files_built_ph): Use the GNU
make $(patsubst ...) function instead of $(var:%=foo%bar) if
a simple $(var:.a=.b) is not sufficient.  Apparently Automake
is able to expand some of these constructs statically, but in
some cases it will actually produce the wrong result.  Using
$(patsubst ...) helps to avoid the problem because Automake
does not attempt to expand it.
* glib/glibmm/filelist.am (glibmm_files_built_ph): Use the
$(patsubst ...) function here, too.
* gio/src/filelist.am (giomm_files_used_hg): Define another
intermediate variable to avoid repetition.