Re-enable input file line markers in gmmproc
* tools/pm/WrapParser.pm: Replace hairy uses of scalar(@array)
to get the number of array elements by $#array, which returns
the index of the last element.
(read_file): Re-enable the commented
code which used to insert input file and line information into
the token list.
(parse_and_build_output): Use a simple regular expression to
match the start of a string instead of messing with substr().