Re-enable network IO stuff (Socket, Resolver, etc)
There is a bit of a problem with some of the async stuff that is exhibited by
the included resolver example.  For example, when run the example with the
following arguments:

  ./resolver -c domain.com

There are several critical warnings printed to the terminal, and I have not been
able to debug it to find the cause yet.  I think it must have something to with
the lifetime of the enumerator object during the next_async() calls and
sigc::bind, etc. but I can't figure it out yet.