AsyncResult: Add get_source_object_base(), deprecating get_source_object().
* gio/asyncresult.[hg|ccg]: Deprecate get_source_object(), replacing it
with get_source_object_base(), because in giomm, the C++ wrapper object
might be a Glib::Interface, but not a Glib::Object (though the underlying
C instance must be a GObject). This happens if giomm does not know about
the GType of the underlying C Object, which is a legitimate situation.
* tests/giomm_asyncresult_sourceobject/main.cc: Added this test case
from Michael Hasselmann, from bug #608269.