Fix missing doc-install.pl problem with the tarball build.
* configure.ac: Remove the call to MM_CONFIG_DOCTOOL_DIR() so we do not copy
the mm-common doc utils into glibmm and do not try to use local copies of
them. This is what gtkmm and other modules do so it makes sense to do it now
for glibmm now that glibmm is not the one that supplies these for pkg-config.
This fixes bug #644829 (David Ronis) with the tarball build and
Ubuntu launchpad bug
https://bugs.launchpad.net/ubuntu/+source/glibmm2.4/+bug/735856
(Krzysztof Klimonda)

However, I can see the point of distributing copies in the tarball, which
could make the tarball build not need mm-common at all. But if we want that
then we should decide to do it for gtkmm and other modules too.