Gio::DBus examples: Renamed one and removed one.
* examples/dbus/server.cc: Rename this to server_without_bus.cc to make its
purpose clearer. This is not the common case. It is for private peer-to-peer
connections not involving the usual session or system buses, or any bus
for that matter. Meaning, for instance, that it does not assign names or
provide introspection, I think.
* examples/dbus/peer.cc: Removed this because it seems like just a
complicated version of the server_without_bus example, and this is not in sync
with the C versoin. Well, we should probably add a client for the server
example.