Change taghandlerwriter.pl to parse the gir structure file.
Now generating perl modules is faster and independent from
generated gir files, which may lack some attributes or tags.
Also, it generates API objects and an XML documentation. The latter
is experimental and not used by default though. It was meant to
generate a documentation for gobject-introspection, because the
one there is old and incomplete.