Add _ADD_CONVERSION macro.
It should behave in almost, but not quite, entirely unlike way as old
_CONVERSION macro behaved.

_ADD_CONVERSION takes 5 parameters - from type, to type and next three
parameters describe conversions for 'none', 'container' and 'full'
transfers. Once added, cannot be removed in the file. But may be
overwritten by another _ADD_CONVERSION macro with the same from type
and to type. These conversions are file scoped.