Generate longer names for enum values.
Because of enum breaking the scoping rule, its values has to be unique
in outer scope. So, for now we can't use names from gir files - we
have to use full names without global prefix. There is an option to
use new style enums, that is - enum classes from C++11, which we might
want to use in future. In new style enums, we can just use names from
gir without fear of multiple definitions of the same name.