Gio::Action: Make get_state() templated.
And do the same for get_state_hint(), renaming the originals to
get_state_variant() and get_state_hint_variant(). We had already
decided to break the ABI of these because the method signatures
in the stable release were useless, so this is OK.
We use the same technique elsewhere in glibmm and gtkmm for
Glib::ValueBase.