Glib::Thread: Don't disable more deprecation warnings than necessary
* glib/src/thread.hg: Don't define GLIB_DISABLE_DEPRECATION_WARNINGS in
the .h file, only in the .cc file. If it's defined in the .h file,
it disables warnings in every file that includes glibmm/thread.h or glibmm.h.
Don't include <glibmmconfig.h> in the .cc file. It became unnecessary when
thread.hg was added to glibmm_files_deprecated_hg in glib/src/filelist.am.
Bug #750379.