Visual Studio Projects: Update to 2013
The latest glibmm requires enough C++11 to be supported so that it can be
built, which is Visual Studio 2013 for the Visual Studio compilers.  This
updates the 2010 projects (and their settings) to 2013.  The folder in
which the projects reside in need to be updated, and we would preferably
need to make the projects compatible with Visual Studio 2015, from the 2013
projects, which is not too different from the 2010 ones.