gmmproc: _CLASS_BOXEDTYPE: Make the static swap() noexcept.
This is generally a good idea because it lets standard containers
use the swap() while maintaining their own noexcept.

This would not be an ABI break, hopefully, because this is an inline
method in the header anyway.
We have recently added the noexcept to the member swap() method,
but hopefully that is not an ABI break either. Removing noexcept
would be an ABI break, I think.