ObjectBase: Add initialize_move().
And _move_current_wrapper() for it to use.
Object and Interface can use initialize_move() in their move
constructors. They need to do this because their calls to the
ObjectBase move constructors will not actually run, at least
when in a derived class, because ObjectBase is a virtual base
class (for use in virtual inheritance).
We have initialize() for much the same reason, for the regular
constructors.