tests: Add glibmm_interface_move.
This code got unpleasantly big because I wanted to test a C++
class that wraps a GObject while also implementing a C++ Interface
that wraps a GInterface that the GObject implements. But it seems
to work.

However, the code to test a raw TestInterface is commented out
because the move constructor does not pass the test. I am not sure
if this should even be expected to work:
  Interface interface = std::move(object);
or:
  Base base = std::move(derived);

Surely that would lose the data (and virtual method table) from the
Derived class anyway?