2.45.52
I never released 2.45.1. I had meant to rebase the version change
to be the latest commit  but forgot to do that.
I might not release this one just yet as a tarball either. I just
need a version number for the other *mm projects to check for.
Releasing a tarball might cause disruption if these latest changes
cause problems.