glib/glibmmconfig.h: Allow Build on Visual Studio 2013
Visual Studio 2013 supports enough of C++-11 to support the build of
glibmm, except that it does not support noexcept (which is something
that Visual Studio 2015 supports).  Define noexcept as _NOEXCEPT on
Visual Studio 2013, so that one can build glibmm on it.

Note that since later Visual Studio versions do not allow one to
define known keywords like noexcet as macros, we need to define
_ALLOW_KEYWORD_MACROS to workaround this limitation.

https://bugzilla.gnome.org/show_bug.cgi?id=753271