MSVC Builds: Move MSVC_Net2010 to MSVC_Net2013
This moves all the project files that were in MSVC_Net2010/ to
MSVC_Net2013/, as we can't build glibmm with anything before Visual Studio
2013.  Note that the project files themselves have already been updated
to be Visual Studio 2013-compatible.

Also drop the MSVC_Net2005 and MSVC_Net2008 directories as the projects
there are simply not enough to build glibmm at all.