Update Visual Studio Info in README.win32
The information in there regarding Visual Studio builds needs to be
up-to-date, especially as we dropped support for Visual Studio 2012 and
earlier, so update the info.  It is no longer necessary, nor is it
recommended, to use the /vd2 option with later Visual Studio versions, as
it causes the code to be built incorrectly and causes crashes.