gmmproc: _CONFIGINCLUDE() + _IS_DEPRECATED = #include in .cc file
* tools/m4/base.m4: If both _CONFIGINCLUDE() and _IS_DEPRECATED are used
in the .hg file, then #include <xxxconfig.h> is included before
ifndef xxx_DISABLE_DEPRECATED in the generated .cc file. This is useful if
the mm module is configured with --disable-deprecated-api, because
xxx_DISABLE_DEPRECATED is defined in the config file. The .cc file can be
compiled, but does not generate any executable code. No need to exclude it
from the build by tampering with the filelist.am file.