Gio::SimpleAction: Make set_state() public
This is needed to apply the requested state in a custom handler for
signal_change_state(). It was protected on an assumption that when the C
doc for g_simple_action_set_state() said “This should only be called by
the implementor of the action”, it meant the implementor of a subclass.

But the C docs for signal change-state are clear that it means the
implementor of the signal handler. Without such a user being able to
call set_state(), they can’t apply the new state, unless they subclass
SimpleAction, which is a lot to require for such a simple case as this.

https://bugzilla.gnome.org/show_bug.cgi?id=777953