Visual Studio builds: Require Visual Studio 2017
Visual Studio 2017 is required to support the C++-14 features that
glibmm-2.52 will require, so update the projects to 2017.  Make the
projects look for libsigc++-3.x as well.

Also make sure that the glibmm and giomm projects are up-to-date, and
drop the threading examples projects as these examples have been dropped
from the tree, and ensure the headers "installed" are up-to-date.