ustring: Use std::initializer_list instead of pointer + size
Acrobatics are needed so that the main compose function does not have to
be templated on the number of strings due to the bloat that would cause.

Since C++11, instead of a pointer to an array & its size, we can pass an
std::initializer_list, which pretty much exists for this purpose, as it
uses compiler magic to abstract a variable sized array behind iterators.

The only drawback is that initializer_list does not have an operator[],
but its elements are guaranteed to be allocated in the order defined, so
we can use (*(begin() + index)) – a bit ugly, but we only need it once!

https://bugzilla.gnome.org/show_bug.cgi?id=784211