ustring: Avoid warning due to type confusion
error: comparison of unsigned expression >= 0 is always true
  [-Werror=type-limits]
       if (index >= 0 && index < ilist.size())

digit_value() - N is an int, but ilist.size() is an std::size_t. There’s
rarely a nice way to handle such situations, but here we have an OK one:

As ilist.size() is static_assert()ed in compose() to be in range of int,
we can safely convert to int for use as the bound. For clarity, convert
to a new int, rather than comparing to static_cast<int>(ilist.size()).

Using an int is clearer than e.g. relying on std::size_t underflowing to
its max if digit_value() returns < 0 and hence not being < ilist.size().

https://bugzilla.gnome.org/show_bug.cgi?id=784211