gmmproc: Add new elements types for the docs_override.xml files
* tools/pm/DocsParser.pm: Add handling of substitute_type_name and
substitute_enumerator_name elements in docs_override.xml files.
This is a small part of the fix for enum class and enums inside classes,
discussed in bug 86864, which is implemented in full in the next ABI-
reaking version of glibmm. This particular part does not break ABI,
it just improves the generated documentation.