gmmproc: Warn if parameter lists are not compatible
* tools/pm/DocsParser.pm: append_parameter_docs(): Warn if the lengths of
the parameter lists in the docs.xml file, the methods.defs file and the _WRAP
macro are not compatible. Don't assume that the last parameter is always a
GError** output parameter just because it's called "error".
Without this fix a too long parameter list in the docs.xml file caused very
cryptic error messages, a too short list usually went undetected, but the
generated documentation was incomplete.