build: Add NMake Makefiles
This adds a set of NMake Makefiles which can be used to build glibmm
with Visual Studio 2017 and later.  This will supersede the Visual
Studio 2017 project files, as this approach will reduce the likelihood
of the Visual Studio build files becoming out-of-date as this also
reads from the various filelist.am's under glib/ and gio/.

The existing gendef, and generated [glib|gio]mmconfig.h and generated
[glib|gio]mm.rc will continue to be used.

The NMake Makefiles will now be distributed instead of the Visual Studio
project files from this point on.

The Visual Studio project files will be removed in the next commit