build: Add NMake Makefiles
This adds a set of NMake Makefiles that can be used to build glibmm with
Visual Studio 2015 or later (2013 is not supported as it does not have
enough C++-11 support).  These NMake Makefiles are now dist'ed instead
of the Visual Studio 2015 projects as it is intended that the NMake
Makefiles will replace the Visual Studio projects, for maintenance
reasons.

Note that for the C++11 versions of glibmm, we name the DLLs and LIBs in
the form of <library>-vc140-2_4.[dll|lib] or
<library>-vc140-d-2_4.[dll|lib] since for both Visual Studio 2015 and
2017 builds as they both link to the Visual Studio v140 C/C++ runtimes.