tests/glibmm_ustring_format: Test output, iomanips
We only tested here whether single arguments got formatted without
throwing exceptions, but not whether they got formatted into the
expected output or combinations of multiple types and I/O manipulators.

This adds a test for that, so we at least have something to ensure that
things are working correctly, e.g. if refactoring format() as I just did