ustring: Avoid un(necessary|clear) ptr @ Stringify
Rather than storing a pointer in the temporary Stringify objects, we can
instead store a reference, which is clearer about it being non-optional
and makes for nicer syntax. Then we return that and use operator& when
forwarding it to compose_private() in the std::initializer_list, since
containers can't hold references, so we do need a pointer there. Unless:

We could perhaps go a step further and make that an initializer_list of
std::reference_wrapper, avoiding pointers altogether while probably not
incurring overhead as it is a simple wrapper, but that's one for later.