Change the ABI to glibmm-2.60
So we can use the 2.58 version numbers for stable releases in the
glibmm-2.4 ABI series.
We don't need to release a stable ABI-parallel glibmm until
we need to release gtkmm 4.0, and that won't happen until GTK+ 4.0.0
happens, and we don't know when that might be.

We've done similar ABI name changes several times before.