fileutils|IOChannel: Don’t use @retval on out args
as it is for documenting particular values of the return variable.

Replace @retval with @param[out], and while here, reorder some arguments
in doc comments to match the order they have in function argument lists.

See https://gitlab.gnome.org/GNOME/gtkmm/merge_requests/9 for more
disussion on the rationale for this.