Gio::ThemedIcon: Add create(const std::vector<Glib::ustring>& iconnames)
And change the type of icon names to Glib::ustring in methods where it
was std::string. Gtkmm consistently uses Glib::ustring for icon names,
but Gio::ThemedIcon has been inconsistent. (The string returned by
Gio::Icon::to_string() is not an icon name. It shall remain a std::string.)