gmmproc: _WRAP_PROPERTY: Check the data type
* glib/glibmm/value_custom.h: Add template class
Glib::Traits::ValueCompatibleWithWrapProperty<> that checks if the template
parameter names a type that can be used with _WRAP_PROPERTY
and _WRAP_CHILD_PROPERTY.
* tools/m4/property.m4:
* tools/pm/Output.pm:
* tools/pm/WrapParser.pm: _WRAP_PROPERTY and _WRAP_CHILD_PROPERTY generate
a static_assert() that checks if the data type is acceptable. The generation
of the static_assert() can be suppressed with the new no_type_check parameter.