ustring: Fix warning/errors if sprintf("fmt only")
If no args to be substituted are given, there is nothing to do, so the
fmt string is returned as-is without substitution. This is an obvious
case of mismatched format/args that we can check. Not doing so causes
warnings/errors with common compiler options, as it is a security risk.

For instance, -Wformat-security causes GCC to complain thusly:
> format not a string literal and no format arguments

Not passing arguments but passing a variable string that might contain
placeholders and thus instruct the compiler to try grabbing random
unrelated things off the stacks is obviously not something we want to do
– nor do we want to do it any *other* time, but this is the one case we
can obviously catch ourselves, rather than hoping any compiler does.

The argument for not just static_assert()ing that sizeof...(args) >= 1
is that this might conceivably be used in generic code where doing the
former would be annoying, and it's easy enough to make that work anyway.

https://gitlab.gnome.org/GNOME/glibmm/issues/21#note_537551