Gio::SocketControlMessage: Add deserialize_vfunc_callback()
The deserialize vfunc in GLib is a class virtual function (not associated
with an instance). Such functions don't exist in C++. But it must be wrapped
in one way or another. g_socket_control_message_deserialize() assumes that
all subclasses of GSocketControlMessage override this vfunc. A user-program
can crash, if any subclass does not.

This patch that does not add API. A more complete fix can be committed,
when API can be added. Fixes #52