Glib::Property: Update for compatibility with Gtk::Builder
When an object is created by GtkBuilder, the GObject-derived C object is
created and its properties set before there is a C++ wrapper to store
the property values in. Glib::custom_set_property_callback() stores
property values in a data structure reached via a GQuark in the object.
PropertyBase::lookup_property() copies those property values to the
PropertyBase objects when the C++ wrapper is created.