Binding: Move SlotTypeTransform up and document it
This is more maintainable than all the arguments to all the overloads of
bind_property() manually repeating the signature in their documentation,
plus more user-friendly as we document it, users can use the typedef, &c

Note: My next thought was we can just take SlotTypeTransforms in the
argument list, instead of totally arbitrary functors, so the conversion
would be done at call site instead of via extra code in the body. But I
didn't get that working yet, and it should be a separate commit anyway.