MSVC_NMake: Add rules to generate and build sources
...from a GIT checkout, from the various .ccg/.hg and the *.[cc|h].m4 sources,
as well as the wrap_init.cc sources.  This will obviously require PERL, as well
as a Cygwin or MSYS installation that has a working m4 executable for Windows
for this to work.  Note that release tarballs continue to build on Visual
Studio as they did before, without the need for PERL nor m4.

Note that this does not yet support generating [glib|gio]mmconfig.h nor
[glib|gio]mm.rc from their .in counterparts, as well as tools\gmmproc
and tools\generate_wrap_init.pl, which will be done in a later commit.