meson: Build items in tools/extra_defs_gen on MSVC too
Ensure we can generate the import lib for glibmm_generate_extra_defs on
Visual Studio builds by using the same technique we used for glibmm and
giomm, and build the two programs that is to be built along side with
it.

Also re-order the items in meson.build a bit so that we are able to use
our gendef tool.