NMake Makefiles: Clean up building tests and examples
Instead of hand-written rules, just make use of what we have to generate the
rules on the fly for most of the tests and examples, since the rules are more
or less the same for each of the test and example programs.  We only need to
single-out the ones that aren't buildable under Windows and those that have
multiple programs under each directory, as opposed to one example/test program
per directory.

Even for those directories that have multiple example programs, we can pretty
easily generate the rules for them.

Also remove repeated parts that were added by accident.

This can shrink the NMake Makefiles by a bit.