NMake Makefiles: Use Toolset version in DLL names by default
This updates the build to use the toolset version in the final DLL
and .lib filenames, similar to what is done in Boost, instead of
the Visual Studio version.

This means that by default, we will use 'vc141' instead of 'vc150'
in Visual Studio 2017 builds, and 'vc142' instead of 'vc160' in
Visual Studio 2019 builds.

If using the former naming is desired, use 'USE_COMPAT_LIBS=1' in
the NMake command line, albeit all such binaries will now be named
with 'vc150' in the DLL and library bames